The import statement inside the package's __init__.py ( from xbmcjson import * ) doesn't work under Python 3. In Python3 the relative path must be explicitly stated.
Before the correction, a test app trying to use the packages with ( from xbmcjson import XBMC ) would trigger an ImportError exception (cannot import name 'XBMC') .
If the app tried to import the whole package ( from xbmcjson import * ), the import would seem to succeed but the first use of the XBMC symbol would trigger a NameError exception (name 'XBMC' is not defined)

I added a tiny dot in front of xbmcjson ( from .xbmcjson import * ) to make it more explicit and that seems to fix both problems.

It has been tested on Linux under Python 3.4.2 and Python 2.7.8 and it seems to work under both versions.